test out FrontendAction and BackendAction
