--- django_unicorn/actions/__init__.py
@@ -1,10 +1,0 @@
-
-# order of imports is important to prevent circular deps
-from .base import Action, ActionResult
-
-from .call_method import CallMethod
-from .refresh import Refresh
-from .reset import Reset
-from .sync_input import SyncInput
-from .toggle import Toggle
-from .validate import Validate
