+++ django_unicorn/actions/backend/reset.py
@@ -1,17 +1,18 @@
-from django_unicorn.components import Component
 from django_unicorn.actions.base import Action, ActionResult
+from django_unicorn.components import Component
+
 
 class Reset(Action):
-    
+
     action_type = "callMethod"
     method_name = "$reset"
-    
+
     def apply(
-        self, 
-        component: Component, 
+        self,
+        component: Component,
         request, # : ComponentRequest,
     ) -> ActionResult:
-        
+
         # create a clean object
         updated_component = Component.create(
             # we keep the original component's id and name
@@ -23,5 +24,5 @@ class Reset(Action):
 
         #  Explicitly remove all errors and prevent validation from firing before render()
         updated_component.errors = {}
-        
+
         return updated_component
