+++ django_unicorn/actions/backend/sync_input.py
@@ -1,13 +1,14 @@
-from django_unicorn.components import Component
 from django_unicorn.actions.base import Action, ActionResult
+from django_unicorn.components import Component
+
 
 class SyncInput(Action):
-    
+
     action_type = "syncInput"
-    
+
     def apply(
-        self, 
-        component: Component, 
+        self,
+        component: Component,
         request, # : ComponentRequest,
     ) -> ActionResult:
         raise NotImplementedError()
