+++ django_unicorn/actions/frontend/__init__.py
@@ -1,0 +1,3 @@
+
+# order of imports is important to prevent circular deps
+from .base import FrontendAction
