+++ django_unicorn/actions/frontend/hash_update.py
@@ -1,0 +1,12 @@
+
+class HashUpdate(Update):
+    """
+    Updates the current URL hash from an action method.
+    """
+
+    def __init__(self, url_hash: str):
+        """
+        Args:
+            param url_hash: The URL hash to change. Example: `#model-123`.
+        """
+        self.hash = url_hash
