+++ django_unicorn/actions/frontend/redirect.py
@@ -1,0 +1,1 @@
+
